requirements.txt
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.1
defusedxml==0.7.1
Django==4.2.17
django-filter==25.1
django-redis==5.4.0
django-rest-swagger==2.2.0
django-social-auth==0.7.28
djangorestframework==3.15.2
djangorestframework-simplejwt==5.1.0
drf-yasg==1.21.8
httplib2==0.22.0
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
oauth2==1.9.0.post1
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.2
pillow==11.1.0
psycopg==3.2.4
psycopg-binary==3.2.4
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.1
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
setuptools==72.1.0
simplejson==3.20.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wheel==0.43.0
gunicorn
celery